@@ -1,9 +1,18 @@
 """
 Client for the Chat service API.
 """
-from typing import Dict, Any, Optional
+from typing import Optional
 
-from .models import MessageRequestModel, CreateChatRequest
+from .models import (
+    MessageRequest,
+    CreateChatRequest,
+    ChatListResponse,
+    MessageResponse,
+    CreateChatResponse,
+    GetMessagesRequest,
+    GetMessagesResponse,
+    GetChatsRequest
+)
 from ..base_client import BaseClient
 
 
@@ -12,75 +21,69 @@ class ChatService(BaseClient):
 
     def __init__(self, **kwargs):
         """Initialize the chat service client."""
-        super().__init__(service_name="chat", **kwargs)
+        super().__init__(service="chat", **kwargs)
 
-    async def send_message(
+    async def create_message(
             self,
-            request: MessageRequestModel,
-            user_agent: str,
-            x_client_info: str,
-            admin_token: Optional[str] = None
-    ) -> Dict[str, Any]:
+            request: MessageRequest,
+            user_agent: Optional[str] = None,  # Just for Basalam internal team usage!!!
+            x_client_info: Optional[str] = None,  # Just for Basalam internal team usage!!!
+    ) -> MessageResponse:
         """
-        Send a message.
+        Create a message.
 
         Args:
             request: The message request model.
             user_agent: The User-Agent header value.
             x_client_info: The X-Client-Info header value.
-            admin_token: Optional Admin-Token header value.
 
         Returns:
-            The response from the API.
+            MessageResponse: The response from the API.
         """
-        endpoint = "/v2/message"
-        headers = {
-            "User-Agent": user_agent,
-            "X-Client-Info": x_client_info
-        }
-        if admin_token:
-            headers["Admin-Token"] = admin_token
+        endpoint = "/v3/messages"
+        headers = {}
+        if user_agent:
+            headers["User-Agent"] = user_agent
+        if x_client_info:
+            headers["X-Client-Info"] = x_client_info
 
-        response = await self._post(endpoint, json=request.dict(), headers=headers)
-        return response
+        response = await self._post(endpoint, json_data=request.model_dump(exclude_none=True), headers=headers)
+        return MessageResponse(**response)
 
-    def send_message_sync(
+    def create_message_sync(
             self,
-            request: MessageRequestModel,
-            user_agent: str,
-            x_client_info: str,
-            admin_token: Optional[str] = None
-    ) -> Dict[str, Any]:
+            request: MessageRequest,
+            user_agent: Optional[str] = None,  # Just for Basalam internal team usage!!!
+            x_client_info: Optional[str] = None,  # Just for Basalam internal team usage!!!
+    ) -> MessageResponse:
         """
-        Send a message (synchronous version).
+        Create a message (synchronous version).
 
         Args:
             request: The message request model.
             user_agent: The User-Agent header value.
             x_client_info: The X-Client-Info header value.
-            admin_token: Optional Admin-Token header value.
 
         Returns:
-            The response from the API.
+            MessageResponse: The response from the API.
         """
-        endpoint = "/v2/message"
-        headers = {
-            "User-Agent": user_agent,
-            "X-Client-Info": x_client_info
-        }
-        if admin_token:
-            headers["Admin-Token"] = admin_token
+        endpoint = "/v3/messages"
+        headers = {}
+        if user_agent:
+            headers["User-Agent"] = user_agent
+        if x_client_info:
+            headers["X-Client-Info"] = x_client_info
 
-        response = self._post_sync(endpoint, json=request.dict(), headers=headers)
-        return response
+        response = self._post_sync(endpoint, json_data=request.model_dump(exclude_none=True), headers=headers)
+        return MessageResponse(**response)
 
     async def create_chat(
             self,
             request: CreateChatRequest,
-            x_creation_tags: Optional[str] = None,
-            x_user_session: Optional[str] = None,
-            x_client_info: Optional[str] = None
-    ) -> Dict[str, Any]:
+            x_creation_tags: Optional[str] = None,  # Just for Basalam internal team usage!!!
+            x_user_session: Optional[str] = None,  # Just for Basalam internal team usage!!!
+            x_client_info: Optional[str] = None  # Just for Basalam internal team usage!!!
+    ) -> CreateChatResponse:
         """
         Create a private chat.
 
@@ -91,9 +94,9 @@ class ChatService(BaseClient):
             x_client_info: Optional X-Client-Info header value.
 
         Returns:
-            The response from the API.
+            CreateChatResponse: The response from the API.
         """
-        endpoint = "/v2/chat"
+        endpoint = "/v3/chats"
         headers = {}
         if x_creation_tags:
             headers["X-Creation-Tags"] = x_creation_tags
@@ -102,16 +105,16 @@ class ChatService(BaseClient):
         if x_client_info:
             headers["X-Client-Info"] = x_client_info
 
-        response = await self._post(endpoint, json=request.dict(), headers=headers)
-        return response
+        response = await self._post(endpoint, json_data=request.model_dump(exclude_none=True), headers=headers)
+        return CreateChatResponse(**response)
 
     def create_chat_sync(
             self,
             request: CreateChatRequest,
-            x_creation_tags: Optional[str] = None,
-            x_user_session: Optional[str] = None,
-            x_client_info: Optional[str] = None
-    ) -> Dict[str, Any]:
+            x_creation_tags: Optional[str] = None,  # Just for Basalam internal team usage!!!
+            x_user_session: Optional[str] = None,  # Just for Basalam internal team usage!!!
+            x_client_info: Optional[str] = None  # Just for Basalam internal team usage!!!
+    ) -> CreateChatResponse:
         """
         Create a private chat (synchronous version).
 
@@ -122,9 +125,9 @@ class ChatService(BaseClient):
             x_client_info: Optional X-Client-Info header value.
 
         Returns:
-            The response from the API.
+            CreateChatResponse: The response from the API.
         """
-        endpoint = "/v2/chat"
+        endpoint = "/v3/chats"
         headers = {}
         if x_creation_tags:
             headers["X-Creation-Tags"] = x_creation_tags
@@ -133,85 +136,121 @@ class ChatService(BaseClient):
         if x_client_info:
             headers["X-Client-Info"] = x_client_info
 
-        response = self._post_sync(endpoint, json=request.dict(), headers=headers)
-        return response
+        response = self._post_sync(endpoint, json_data=request.model_dump(exclude_none=True), headers=headers)
+        return CreateChatResponse(**response)
 
     async def get_messages(
             self,
-            chat_id: int,
-            msg_id: Optional[int] = None,
-            limit: int = 20,
-            chat_type: str = "ALL",
-            order: str = "DESC",
-            op: str = "<",
-            temp_id: Optional[int] = None
-    ) -> Dict[str, Any]:
+            request: GetMessagesRequest,
+    ) -> GetMessagesResponse:
         """
         Get messages from a chat.
 
         Args:
-            chat_id: The ID of the chat.
-            msg_id: Optional message ID to start from.
-            limit: Maximum number of messages to return.
-            chat_type: Type of chat messages to return.
-            order: Order of messages (ASC or DESC).
-            op: Operator for message ID comparison.
-            temp_id: Optional temporary message ID.
+            request: The get messages request model containing chat_id and query parameters.
 
         Returns:
-            The response from the API.
+            GetMessagesResponse: The response containing the list of messages.
         """
-        endpoint = f"/v2/chat/{chat_id}/messages"
+        endpoint = f"/v3/messages"
         params = {
-            "limit": limit,
-            "chatType": chat_type,
-            "order": order,
-            "op": op
+            "chat_id": request.chat_id,
+            "limit": request.limit,
+            "order": request.order,
+            "cmp": request.cmp
         }
-        if msg_id is not None:
-            params["msgId"] = msg_id
-        if temp_id is not None:
-            params["temp_id"] = temp_id
+        if request.message_id is not None:
+            params["message_id"] = request.message_id
 
         response = await self._get(endpoint, params=params)
-        return response
+        return GetMessagesResponse(**response)
 
     def get_messages_sync(
             self,
-            chat_id: int,
-            msg_id: Optional[int] = None,
-            limit: int = 20,
-            chat_type: str = "ALL",
-            order: str = "DESC",
-            op: str = "<",
-            temp_id: Optional[int] = None
-    ) -> Dict[str, Any]:
+            request: GetMessagesRequest,
+    ) -> GetMessagesResponse:
         """
         Get messages from a chat (synchronous version).
 
         Args:
-            chat_id: The ID of the chat.
-            msg_id: Optional message ID to start from.
-            limit: Maximum number of messages to return.
-            chat_type: Type of chat messages to return.
-            order: Order of messages (ASC or DESC).
-            op: Operator for message ID comparison.
-            temp_id: Optional temporary message ID.
+            request: The get messages request model containing chat_id and query parameters.
 
         Returns:
-            The response from the API.
+            GetMessagesResponse: The response containing the list of messages.
         """
-        endpoint = f"/v2/chat/{chat_id}/messages"
+        endpoint = f"/v3/messages"
         params = {
-            "limit": limit,
-            "chatType": chat_type,
-            "order": order,
-            "op": op
+            "chat_id": request.chat_id,
+            "limit": request.limit,
+            "order": request.order,
+            "cmp": request.cmp
         }
-        if msg_id is not None:
-            params["msgId"] = msg_id
-        if temp_id is not None:
-            params["temp_id"] = temp_id
+        if request.message_id is not None:
+            params["message_id"] = request.message_id
 
         response = self._get_sync(endpoint, params=params)
-        return response
+        return GetMessagesResponse(**response)
+
+    async def get_chats(
+            self,
+            request: GetChatsRequest,
+    ) -> ChatListResponse:
+        """
+        Get list of chats.
+
+        Args:
+            request: The get chats request model containing query parameters.
+
+        Returns:
+            ChatListResponse: The list of chats based on OpenAPI specification.
+        """
+        endpoint = f"/v3/chats"
+        params = {
+            "limit": request.limit,
+            "order_by": request.order_by.value
+        }
+        if request.updated_from is not None:
+            params["updated_from"] = request.updated_from
+        if request.updated_before is not None:
+            params["updated_before"] = request.updated_before
+        if request.modified_from is not None:
+            params["modified_from"] = request.modified_from
+        if request.modified_before is not None:
+            params["modified_before"] = request.modified_before
+        if request.filters is not None:
+            params["filters"] = request.filters.value
+
+        response = await self._get(endpoint, params=params)
+        return ChatListResponse(**response)
+
+    def get_chats_sync(
+            self,
+            request: GetChatsRequest,
+    ) -> ChatListResponse:
+        """
+        Get list of chats (synchronous version).
+
+        Args:
+            request: The get chats request model containing query parameters.
+
+        Returns:
+            ChatListResponse: The list of chats based on OpenAPI specification.
+        """
+        endpoint = f"/v3/chats"
+        params = {
+            "limit": request.limit,
+            "order_by": request.order_by.value
+        }
+        if request.updated_from is not None:
+            params["updated_from"] = request.updated_from
+        if request.updated_before is not None:
+            params["updated_before"] = request.updated_before
+        if request.modified_from is not None:
+            params["modified_from"] = request.modified_from
+        if request.modified_before is not None:
+            params["modified_before"] = request.modified_before
+        if request.filters is not None:
+            params["filters"] = request.filters.value
+
+        response = self._get_sync(endpoint, params=params)
+        return ChatListResponse(**response)
