@@ -4,7 +4,7 @@ Basalam Python SDK for accessing the Basalam API.
 This package provides a clean and simple interface to interact with Basalam's microservices.
 """
 
-from .auth import BaseAuth, ClientCredentials, AuthorizationCode, Scope
+from .auth import BaseAuth, ClientCredentials, AuthorizationCode, PersonalToken, Scope
 from .basalam_client import BasalamClient
 from .base_client import BaseClient
 from .chat.client import ChatService
@@ -16,10 +16,9 @@ from .order.client import OrderService
 from .order_processing.client import OrderProcessingService
 from .search.client import SearchService
 from .upload.client import UploadService
+from .version import __version__, __sdk_name__, get_user_agent
 from .wallet.client import WalletService
 from .webhook.client import WebhookService
-
-__version__ = "0.1.0"
 
 __all__ = [
     # Main client
@@ -30,6 +29,7 @@ __all__ = [
     "BaseAuth",
     "ClientCredentials",
     "AuthorizationCode",
+    "PersonalToken",
     "Scope",
 
     # Configuration
@@ -50,4 +50,9 @@ __all__ = [
     "UploadService",
     "ChatService",
     "WebhookService",
+
+    # Version and utilities
+    "__version__",
+    "__sdk_name__",
+    "get_user_agent",
 ]
