@@ -3,15 +3,206 @@ Core service module for the Basalam SDK.
 """
 from .client import CoreService
 from .models import (
-    Product,
-    ProductStatus,
-    ProductImage,
-    ProductPrice,
-    ProductAttribute,
-    Category,
-    CategoryType,
-    User,
-    UserRole,
-    Address,
-    UserInfo,
+    ProductStatusInputEnum,
+    AttributeGroupResponseSchema,
+    AttributesResponse,
+    BulkActionItem,
+    BulkProductUpdateItemResponse,
+    BulkProductsUpdateRequestSchema,
+    BulkProductsUpdateResponseSchema,
+    BulkProductsUpdatesCountResponse,
+    BulkProductsUpdatesListResponse,
+    CategoriesResponse,
+    CategoryListItemResponseSchema,
+    CategoryProductResponse,
+    CategoryResponse,
+    CategoryTypeResponse,
+    CategoryUnitTypeResponse,
+    ChangeUserMobileConfirmSchema,
+    ChangeUserMobileRequestSchema,
+    ChangeVendorMobileConfirmSchema,
+    ChangeVendorMobileRequestSchema,
+    CityResponse,
+    ConfirmCurrentUserMobileConfirmSchema,
+    CreateDiscountRequestSchema,
+    CreateVendorSchema,
+    DeleteDiscountRequestSchema,
+    DiscountProductFilterSchema,
+    FreeShippingResponseSchema,
+    GenderResponse,
+    GetProductsQuerySchema,
+    GetVendorProductsSchema,
+    HomeTabSettingsResponse,
+    IllegalPhotoResponse,
+    ImageResponse,
+    LocationDeploymentResponseSchema,
+    MarkedTypeResponse,
+    NavigationResponseSchema,
+    PackagingDimensionsRequestItem,
+    PackagingDimensionsResponseSchema,
+    PrivateUserResponse,
+    PrivateVendorResponse,
+    ProductAttributeRequestItem,
+    ProductAttributeResponse,
+    ProductAttributeValueResponse,
+    ProductFilterSchema,
+    ProductItemResponse,
+    ProductListResponse,
+    ProductRequestSchema,
+    ProductResponseSchema,
+    ProductRevisionDataResponse,
+    ProductRevisionResponse,
+    ProductShelfResponse,
+    ProductSortTypeResponse,
+    ProductVariantResponse,
+    PropertyResponse,
+    PropertyValueResponse,
+    ProvinceResponse,
+    PublicUserResponse,
+    PublicVendorResponse,
+    PublicVendorSimpleResponse,
+    RangeFilterItem,
+    ReferralJourneyResponse,
+    RejectionReasonResponse,
+    ResultResponse,
+    RevisionMetadataResponse,
+    ShippingDataRequestItem,
+    ShippingDataResponse,
+    ShippingMethodInfo,
+    ShippingMethodItemResponse,
+    ShippingMethodListResponse,
+    ShippingMethodResponse,
+    ShippingMethodUpdateItem,
+    StatusResponse,
+    UnitTypeResponse,
+    UnsuccessfulBulkUpdateProducts,
+    UnsuccessfulProductItem,
+    UpdateProductRequestItem,
+    UpdateProductResponseItem,
+    UpdateProductVariationSchema,
+    BatchUpdateProductsRequest,
+    UpdateShippingMethodSchema,
+    UpdateUserBankInformationSchema,
+    UpdateVariantRequestItem,
+    UpdateVendorSchema,
+    UpdateVendorStatusResponse,
+    UpdateVendorStatusSchema,
+    UserCardsOtpSchema,
+    UserCardsSchema,
+    UserVerificationSchema,
+    UserVerifyBankInformationSchema,
+    VariantPropertyRequestItem,
+    VariantPropertyResponse,
+    VariantRequestItem,
+    VendorLegalDataSchema,
+    VendorLegalRequestSchema,
+    VendorSettingResponse,
+    VideoDetailResponse,
+    ProductBulkActionTypeEnum,
+    ProductBulkFieldInputEnum,
 )
+
+__all__ = [
+    # Client
+    "CoreService",
+
+    # Enums
+    "ProductStatusInputEnum",
+    "ProductBulkActionTypeEnum",
+    "ProductBulkFieldInputEnum",
+    # Models (alphabetical order)
+    "AttributeGroupResponseSchema",
+    "AttributesResponse",
+    "BulkActionItem",
+    "BulkProductUpdateItemResponse",
+    "BulkProductsUpdateRequestSchema",
+    "BulkProductsUpdateResponseSchema",
+    "BulkProductsUpdatesCountResponse",
+    "BulkProductsUpdatesListResponse",
+    "CategoriesResponse",
+    "CategoryListItemResponseSchema",
+    "CategoryProductResponse",
+    "CategoryResponse",
+    "CategoryTypeResponse",
+    "CategoryUnitTypeResponse",
+    "ChangeUserMobileConfirmSchema",
+    "ChangeUserMobileRequestSchema",
+    "ChangeVendorMobileConfirmSchema",
+    "ChangeVendorMobileRequestSchema",
+    "CityResponse",
+    "ConfirmCurrentUserMobileConfirmSchema",
+    "CreateDiscountRequestSchema",
+    "CreateVendorSchema",
+    "DeleteDiscountRequestSchema",
+    "DiscountProductFilterSchema",
+    "FreeShippingResponseSchema",
+    "GenderResponse",
+    "GetProductsQuerySchema",
+    "GetVendorProductsSchema",
+    "HomeTabSettingsResponse",
+    "IllegalPhotoResponse",
+    "ImageResponse",
+    "LocationDeploymentResponseSchema",
+    "MarkedTypeResponse",
+    "NavigationResponseSchema",
+    "PackagingDimensionsRequestItem",
+    "PackagingDimensionsResponseSchema",
+    "PrivateUserResponse",
+    "PrivateVendorResponse",
+    "ProductAttributeRequestItem",
+    "ProductAttributeResponse",
+    "ProductAttributeValueResponse",
+    "ProductFilterSchema",
+    "ProductItemResponse",
+    "ProductListResponse",
+    "ProductRequestSchema",
+    "ProductResponseSchema",
+    "ProductRevisionDataResponse",
+    "ProductRevisionResponse",
+    "ProductShelfResponse",
+    "ProductSortTypeResponse",
+    "ProductVariantResponse",
+    "PropertyResponse",
+    "PropertyValueResponse",
+    "ProvinceResponse",
+    "PublicUserResponse",
+    "PublicVendorResponse",
+    "PublicVendorSimpleResponse",
+    "RangeFilterItem",
+    "ReferralJourneyResponse",
+    "RejectionReasonResponse",
+    "ResultResponse",
+    "RevisionMetadataResponse",
+    "ShippingDataRequestItem",
+    "ShippingDataResponse",
+    "ShippingMethodInfo",
+    "ShippingMethodItemResponse",
+    "ShippingMethodListResponse",
+    "ShippingMethodResponse",
+    "ShippingMethodUpdateItem",
+    "StatusResponse",
+    "UnitTypeResponse",
+    "UnsuccessfulBulkUpdateProducts",
+    "UnsuccessfulProductItem",
+    "UpdateProductRequestItem",
+    "UpdateProductResponseItem",
+    "UpdateProductVariationSchema",
+    "BatchUpdateProductsRequest",
+    "UpdateShippingMethodSchema",
+    "UpdateUserBankInformationSchema",
+    "UpdateVariantRequestItem",
+    "UpdateVendorSchema",
+    "UpdateVendorStatusResponse",
+    "UpdateVendorStatusSchema",
+    "UserCardsOtpSchema",
+    "UserCardsSchema",
+    "UserVerificationSchema",
+    "UserVerifyBankInformationSchema",
+    "VariantPropertyRequestItem",
+    "VariantPropertyResponse",
+    "VariantRequestItem",
+    "VendorLegalDataSchema",
+    "VendorLegalRequestSchema",
+    "VendorSettingResponse",
+    "VideoDetailResponse",
+]
