@@ -1,197 +1,46 @@
 """
-Data models for the Basalam Order Service.
+Data models for the Basalam OrderEnum Service.
 
 This module contains all the data models used by the order service client.
 """
 
-from datetime import datetime
 from enum import Enum
-from typing import Dict, List, Optional, Any, Union
+from typing import Dict, List, Optional, Any
 
 from pydantic import BaseModel
 
 
-class OrderStatusEnum(str, Enum):
-    """Order status enumeration."""
-    PENDING = "pending"
-    CONFIRMED = "confirmed"
-    PROCESSING = "processing"
-    SHIPPED = "shipped"
-    DELIVERED = "delivered"
-    CANCELLED = "cancelled"
-    REFUNDED = "refunded"
+class City(BaseModel):
+    """City model."""
+    id: Optional[int] = None
+    title: Optional[str] = None
+    parent: Optional['City'] = None
 
 
-class PaymentStatusEnum(str, Enum):
-    """Payment status enumeration."""
-    PENDING = "pending"
-    PAID = "paid"
-    FAILED = "failed"
-    REFUNDED = "refunded"
-
-
-class PaymentMethodEnum(str, Enum):
-    """Payment method enumeration."""
-    ONLINE = "online"
-    CASH_ON_DELIVERY = "cash_on_delivery"
-    WALLET = "wallet"
-
-
-class Address(BaseModel):
-    """Address information."""
-    id: int
-    user_id: int
-    title: str
-    receiver_name: str
-    receiver_phone: str
-    province: str
-    city: str
-    address: str
-    postal_code: str
-    latitude: Optional[float] = None
-    longitude: Optional[float] = None
-    is_default: bool = False
-    created_at: datetime = datetime.now()
-    updated_at: datetime = datetime.now()
-
-
-class OrderItem(BaseModel):
-    """Order item information."""
-    id: int
-    order_id: int
-    product_id: int
-    variant_id: Optional[int] = None
-    quantity: int
-    unit_price: int
-    total_price: int
-    discount_amount: int = 0
-    final_price: int = 0
-    product_name: str = ""
-    product_image: Optional[str] = None
-    created_at: datetime = datetime.now()
-    updated_at: datetime = datetime.now()
-
-
-class Order(BaseModel):
-    """Order information."""
-    id: int
-    user_id: int
-    status: OrderStatusEnum
-    payment_status: PaymentStatusEnum
-    payment_method: PaymentMethodEnum
-    total_amount: int
-    shipping_amount: int
-    discount_amount: int = 0
-    final_amount: int = 0
-    address: Address
-    items: List[OrderItem] = []
-    tracking_code: Optional[str] = None
-    created_at: datetime = datetime.now()
-    updated_at: datetime = datetime.now()
-
-
-class OrderCreateRequest(BaseModel):
-    """Request model for creating an order."""
-    address_id: int
-    payment_method: PaymentMethodEnum
-    items: List[Dict[str, Any]]
-    coupon_code: Optional[str] = None
-    description: Optional[str] = None
-
-
-class OrderUpdateRequest(BaseModel):
-    """Request model for updating an order."""
-    status: Optional[OrderStatusEnum] = None
-    payment_status: Optional[PaymentStatusEnum] = None
-    tracking_code: Optional[str] = None
-    description: Optional[str] = None
-
-
-class OrderListResponse(BaseModel):
-    """Response model for listing orders."""
-    items: List[Order]
-    total: int
-    page: int
-    per_page: int
-    total_pages: int
-
-
-class OrderDetailResponse(BaseModel):
-    """Response model for order details."""
-    order: Order
-    related_orders: List[Order] = []
-
-
-class OrderCancelRequest(BaseModel):
-    """Request model for cancelling an order."""
-    reason: str
-    description: Optional[str] = None
-
-
-class OrderRefundRequest(BaseModel):
-    """Request model for refunding an order."""
-    reason: str
-    description: Optional[str] = None
-    refund_amount: Optional[int] = None
-
-
-class OrderPaymentRequest(BaseModel):
-    """Request model for processing order payment."""
-    payment_method: PaymentMethodEnum
-    payment_id: Optional[str] = None
-    description: Optional[str] = None
-
-
-class OrderPaymentResponse(BaseModel):
-    """Response model for order payment."""
-    order_id: int
-    payment_id: str
-    payment_url: Optional[str] = None
-    payment_status: PaymentStatusEnum
-    created_at: datetime = datetime.now()
-
-
-class OrderTrackingResponse(BaseModel):
-    """Response model for order tracking."""
-    order_id: int
-    tracking_code: str
-    status: OrderStatusEnum
-    events: List[Dict[str, Any]] = []
-    estimated_delivery: Optional[datetime] = None
-    last_updated: datetime = datetime.now()
-
-
-class ValidationError(BaseModel):
-    """Validation error model."""
-    loc: List[Union[str, int]]
-    msg: str
-    type: str
-
-
-class HTTPValidationError(BaseModel):
-    """HTTP validation error model."""
-    detail: List[ValidationError]
-
-
-class Order(str, Enum):
-    """Order enum."""
+class OrderEnum(str, Enum):
+    """OrderEnum enum."""
     ASC = "ASC"
     DESC = "DESC"
 
 
 class UnpaidInvoiceStatusEnum(str, Enum):
     """Unpaid invoice status enum."""
-    PENDING = "pending"
-    PAID = "paid"
-    FAILED = "failed"
-    REFUNDED = "refunded"
+    SALEABLE = "saleable"
+    PAYABLE = "payable"
+    UNPAID = "unpaid"
+
+
+class PaymentDriver(BaseModel):
+    """Payment driver model."""
+    amount: int
 
 
 class CreatePaymentRequestModel(BaseModel):
     """Create payment request model."""
-    payment_method: str
-    payment_id: Optional[str] = None
-    description: Optional[str] = None
+    pay_drivers: Dict[str, PaymentDriver]
+    callback: str
+    option_code: Optional[str] = None
+    national_id: Optional[str] = None
 
 
 class PaymentCallbackRequestModel(BaseModel):
@@ -206,3 +55,198 @@ class PaymentVerifyRequestModel(BaseModel):
     payment_id: str
     transaction_id: Optional[str] = None
     description: Optional[str] = None
+
+
+# Basket Models - Updated for new structure
+
+class CostBreakdown(BaseModel):
+    """Cost breakdown model."""
+    base: Optional[int] = None
+    discount: Optional[int] = None
+    grand: Optional[int] = None
+
+
+class TotalCostBreakdown(BaseModel):
+    """Total cost breakdown model."""
+    base: Optional[int] = None
+    discount: Optional[int] = None
+    credit: Optional[int] = None
+    bnpl: Optional[int] = None
+    installment: Optional[int] = None
+    pay_lines: Optional[int] = None
+    grand: Optional[int] = None
+
+
+class BasketCosts(BaseModel):
+    """Basket costs model."""
+    delivery: Optional[CostBreakdown] = None
+    products: Optional[CostBreakdown] = None
+    total: Optional[TotalCostBreakdown] = None
+
+
+class BasketAddress(BaseModel):
+    """Basket address model."""
+    id: Optional[int] = None
+    name: Optional[str] = None
+    mobile: Optional[str] = None
+    tel: Optional[str] = None
+    address: Optional[str] = None
+    postal_code: Optional[str] = None
+    is_default: Optional[bool] = None
+    city: Optional[City] = None
+    latitude: Optional[float] = None
+    longitude: Optional[float] = None
+    house_number: Optional[str] = None
+    house_unit: Optional[str] = None
+
+
+class ShippingMethodInfo(BaseModel):
+    """Shipping method information."""
+    id: Optional[int] = None
+    title: Optional[str] = None
+    parent: Optional[Dict[str, Any]] = None
+
+
+class OriginShippingMethod(BaseModel):
+    """Origin shipping method model."""
+    id: Optional[int] = None
+    method: Optional[ShippingMethodInfo] = None
+    delivery_time: Optional[int] = None
+    warehouse: Optional[Dict[str, Any]] = None
+
+
+class OriginParcel(BaseModel):
+    """Origin parcel model."""
+    id: Optional[int] = None
+    shipping_method: Optional[OriginShippingMethod] = None
+    preparation_days: Optional[int] = None
+    delivery_days: Optional[int] = None
+    arrival_delivery_days: Optional[int] = None
+
+
+class Origin(BaseModel):
+    """Origin model."""
+    id: Optional[int] = None
+    title: Optional[str] = None
+    type: Optional[str] = None
+    parcel: Optional[OriginParcel] = None
+    city: Optional[City] = None
+    is_warehouse: Optional[bool] = None
+    vendor_identifiers: Optional[List[int]] = None
+    delivery_costs: Optional[CostBreakdown] = None
+
+
+class BasketProductPhoto(BaseModel):
+    """Basket product photo model."""
+    id: Optional[int] = None
+    original: Optional[str] = None
+    resized: Optional[Dict[str, str]] = None
+
+
+class BasketProductCategory(BaseModel):
+    """Basket product category model."""
+    id: Optional[int] = None
+    title: Optional[str] = None
+
+
+class BasketProduct(BaseModel):
+    """Basket product model."""
+    id: Optional[int] = None
+    title: Optional[str] = None
+    price: Optional[int] = None
+    primary_price: Optional[int] = None
+    stock: Optional[int] = None
+    category: Optional[BasketProductCategory] = None
+    photos: Optional[List[BasketProductPhoto]] = None
+
+
+class BasketVariationProperty(BaseModel):
+    """Basket variation property model."""
+    property: Optional[Dict[str, Any]] = None
+    value: Optional[Dict[str, Any]] = None
+
+
+class BasketVariation(BaseModel):
+    """Basket variation model."""
+    id: Optional[int] = None
+    stock: Optional[int] = None
+    price: Optional[int] = None
+    primary_price: Optional[int] = None
+    properties: Optional[List[BasketVariationProperty]] = None
+
+
+class BasketVendorItem(BaseModel):
+    """Basket vendor item model."""
+    id: Optional[int] = None
+    parcel_id: Optional[int] = None
+    title: Optional[str] = None
+    quantity: Optional[int] = None
+    price: Optional[int] = None
+    primary_price: Optional[int] = None
+    payable_amount: Optional[int] = None
+    delivery_cost: Optional[int] = None
+    total_discount: Optional[int] = None
+    basalam_product_discount: Optional[int] = None
+    basalam_delivery_discount: Optional[int] = None
+    vendor_product_discount: Optional[int] = None
+    vendor_delivery_discount: Optional[int] = None
+    public_delivery_discount: Optional[int] = None
+    product: Optional[BasketProduct] = None
+    variation: Optional[BasketVariation] = None
+    vendor_coupon: Optional[Dict[str, Any]] = None
+    errors: Optional[List[str]] = None
+    comment: Optional[str] = None
+    is_deleted: Optional[bool] = None
+
+
+class VendorOwnerAvatar(BaseModel):
+    """Vendor owner avatar model."""
+    id: Optional[int] = None
+    original: Optional[str] = None
+    resized: Optional[Dict[str, str]] = None
+
+
+class VendorOwner(BaseModel):
+    """Vendor owner model."""
+    id: Optional[int] = None
+    hash_id: Optional[str] = None
+    name: Optional[str] = None
+    avatar: Optional[VendorOwnerAvatar] = None
+    city: Optional[City] = None
+
+
+class BasketVendor(BaseModel):
+    """Basket vendor model."""
+    id: Optional[int] = None
+    identifier: Optional[str] = None
+    url_alias: Optional[str] = None
+    title: Optional[str] = None
+    logo: Optional[Dict[str, Any]] = None
+    owner: Optional[VendorOwner] = None
+    total_product_amount: Optional[int] = None
+    free_shipping_amount: Optional[int] = None
+    free_shipping_type: Optional[str] = None
+    items: Optional[List[BasketVendorItem]] = None
+    city: Optional[City] = None
+    parcel_identifiers: Optional[List[int]] = None
+    origin_identifiers: Optional[List[int]] = None
+    preparation_days: Optional[int] = None
+    delivery_days: Optional[int] = None
+    arrival_days: Optional[int] = None
+    delivery_costs: Optional[CostBreakdown] = None
+
+
+class BasketResponse(BaseModel):
+    """Response model for basket endpoint."""
+    id: Optional[int] = None
+    item_count: Optional[int] = None
+    show_recipient_mobile: Optional[bool] = None
+    delivery_method: Optional[str] = None
+    costs: Optional[BasketCosts] = None
+    errors: Optional[List[str]] = None
+    error_count: Optional[int] = None
+    coupon: Optional[Dict[str, Any]] = None
+    option_code: Optional[str] = None
+    address: Optional[BasketAddress] = None
+    origins: Optional[List[Origin]] = None
+    vendors: Optional[List[BasketVendor]] = None
