@@ -1,5 +1,5 @@
 """
-Order Processing service module for the Basalam SDK.
+OrderEnum Processing service module for the Basalam SDK.
 
 This module provides access to Basalam's order processing service APIs.
 """
@@ -8,20 +8,28 @@ from .client import OrderProcessingService
 from .models import (
     ResourceStats,
     ParcelStatus,
-    PaginationMeta,
     FileResponse,
-    CategoryResponse,
-    ProductSummaryResponse,
-    ProductVariationResponse,
-    ItemStatusResponse,
-    Item,
-    ItemSummary,
-    VendorResponse,
+    AmountDrivers,
+    City,
+    User,
+    Recipient,
+    Customer,
+    Status,
+    ShippingMethodOption,
+    ShippingMethod,
+    Vendor,
+    Product,
+    Property,
+    PropertyValue,
+    VariationProperty,
+    Variation,
+    ParcelItem,
+    Parcel,
+    Order,
     ParcelResponse,
-    OrderResponse,
+    ParcelSummaryResponse,
+    ParcelsResponse,
     OrdersResponse,
-    ItemResponse,
-    ItemsResponse,
     OrderStatsResponse,
     OrderParcelFilter,
     OrderFilter,
@@ -32,20 +40,28 @@ __all__ = [
     "OrderProcessingService",
     "ResourceStats",
     "ParcelStatus",
-    "PaginationMeta",
     "FileResponse",
-    "CategoryResponse",
-    "ProductSummaryResponse",
-    "ProductVariationResponse",
-    "ItemStatusResponse",
-    "Item",
-    "ItemSummary",
-    "VendorResponse",
+    "AmountDrivers",
+    "City",
+    "User",
+    "Recipient",
+    "Customer",
+    "Status",
+    "ShippingMethodOption",
+    "ShippingMethod",
+    "Vendor",
+    "Product",
+    "Property",
+    "PropertyValue",
+    "VariationProperty",
+    "Variation",
+    "ParcelItem",
+    "Parcel",
+    "Order",
     "ParcelResponse",
-    "OrderResponse",
+    "ParcelSummaryResponse",
+    "ParcelsResponse",
     "OrdersResponse",
-    "ItemResponse",
-    "ItemsResponse",
     "OrderStatsResponse",
     "OrderParcelFilter",
     "OrderFilter",
