@@ -5,10 +5,15 @@ This module provides access to Basalam's upload service APIs.
 """
 
 from .client import UploadService
-from .models import FileResponse, UserUploadFileTypeEnum
+from .models import (
+    FileResponse,
+    UserUploadFileTypeEnum,
+    UploadFileRequest,
+)
 
 __all__ = [
     "UploadService",
     "FileResponse",
     "UserUploadFileTypeEnum",
+    "UploadFileRequest",
 ]
