@@ -3,7 +3,7 @@ Client for the Upload service API.
 """
 from typing import Optional, BinaryIO
 
-from .models import FileResponse, UserUploadFileTypeEnum
+from .models import FileResponse, UserUploadFileTypeEnum, UploadFileRequest
 from ..base_client import BaseClient
 
 
@@ -12,7 +12,7 @@ class UploadService(BaseClient):
 
     def __init__(self, **kwargs):
         """Initialize the upload service client."""
-        super().__init__(service_name="upload", **kwargs)
+        super().__init__(service="upload", **kwargs)
 
     async def upload_file(
             self,
@@ -34,14 +34,20 @@ class UploadService(BaseClient):
             The response containing the uploaded file details.
         """
         endpoint = "/v3/files"
+
+        # Prepare files for multipart/form-data
         files = {"file": file}
-        data = {
-            "file_type": file_type
-        }
-        if custom_unique_name is not None:
-            data["custom_unique_name"] = custom_unique_name
-        if expire_minutes is not None:
-            data["expire_minutes"] = expire_minutes
+
+        # Prepare form data according to OpenAPI specification
+        # Create request model to validate parameters, then extract data
+        request = UploadFileRequest(
+            file_type=file_type,
+            custom_unique_name=custom_unique_name,
+            expire_minutes=expire_minutes
+        )
+
+        # Convert to form data, excluding None values
+        data = request.model_dump(exclude_none=True, mode='json')
 
         response = await self._post(endpoint, files=files, data=data)
         return FileResponse(**response)
@@ -66,14 +72,20 @@ class UploadService(BaseClient):
             The response containing the uploaded file details.
         """
         endpoint = "/v3/files"
+
+        # Prepare files for multipart/form-data
         files = {"file": file}
-        data = {
-            "file_type": file_type
-        }
-        if custom_unique_name is not None:
-            data["custom_unique_name"] = custom_unique_name
-        if expire_minutes is not None:
-            data["expire_minutes"] = expire_minutes
+
+        # Prepare form data according to OpenAPI specification
+        # Create request model to validate parameters, then extract data
+        request = UploadFileRequest(
+            file_type=file_type,
+            custom_unique_name=custom_unique_name,
+            expire_minutes=expire_minutes
+        )
+
+        # Convert to form data, excluding None values
+        data = request.model_dump(exclude_none=True, mode='json')
 
         response = self._post_sync(endpoint, files=files, data=data)
         return FileResponse(**response)
