@@ -7,41 +7,47 @@ This module provides access to Basalam's wallet service APIs.
 from .client import WalletService
 from .models import (
     BalanceFilter,
-    BalanceResponse,
-    HistoryPaginationResponse,
-    ReferenceRequestModel,
-    ReferenceTypeEnum,
-    SpendCreditRequest,
-    SpendResponse,
-    RefundRequest,
-    ClientTransferResponse,
-    ClientCreditResponse,
-    CreditResponse,
-    CreditMainTypeEnum,
-    CreditCreationRequest,
+    CanRollbackRefundResponse,
     CreditCreationResponse,
-    RollbackRefundRequest,
+    CreditResponse,
+    CreditTypeResponse,
+    HistoryCreditItemResponse,
+    HistoryItemResponse,
+    HistoryPaginationResponse,
+    HistorySpendItemResponse,
+    HistorySpendResponse,
+    NewHistoryCreditResponse,
     ReasonResponse,
-    DeleteTransferResponse
+    ReferenceRequest,
+    ReferenceResponse,
+    RefundRequest,
+    RollbackRefundRequest,
+    SpendCreditRequest,
+    SpendItemResponse,
+    SpendResponse,
+    SpendSpecificCreditRequest,
 )
 
 __all__ = [
     "WalletService",
     "BalanceFilter",
-    "BalanceResponse",
-    "HistoryPaginationResponse",
-    "ReferenceRequestModel",
-    "ReferenceTypeEnum",
-    "SpendCreditRequest",
-    "SpendResponse",
-    "RefundRequest",
-    "ClientTransferResponse",
-    "ClientCreditResponse",
-    "CreditResponse",
-    "CreditMainTypeEnum",
-    "CreditCreationRequest",
+    "CanRollbackRefundResponse",
     "CreditCreationResponse",
-    "RollbackRefundRequest",
+    "CreditResponse",
+    "CreditTypeResponse",
+    "HistoryCreditItemResponse",
+    "HistoryItemResponse",
+    "HistoryPaginationResponse",
+    "HistorySpendItemResponse",
+    "HistorySpendResponse",
+    "NewHistoryCreditResponse",
     "ReasonResponse",
-    "DeleteTransferResponse"
+    "ReferenceRequest",
+    "ReferenceResponse",
+    "RefundRequest",
+    "RollbackRefundRequest",
+    "SpendCreditRequest",
+    "SpendItemResponse",
+    "SpendResponse",
+    "SpendSpecificCreditRequest",
 ]
