@@ -6,28 +6,46 @@ This module provides access to Basalam's webhook service APIs.
 
 from .client import WebhookService
 from .models import (
-    CreateServiceRequest, ServiceListResource, ServiceResource,
-    CreateWebhookRequest, UpdateWebhookRequest, WebhookResource, WebhookListResource,
-    DeleteWebhookResponse, EventListResource, WebhookLogListResource,
-    RegisterClientRequest, UnRegisterClientRequest, UnRegisterClientResponse,
-    ClientListResource, WebhookRegisteredOnListResource
+    ServiceResource,
+    ServiceListResource,
+    CreateServiceRequest,
+    WebhookResource,
+    WebhookListResource,
+    CreateWebhookRequest,
+    EventResource,
+    EventListResource,
+    UpdateWebhookRequest,
+    DeleteWebhookResponse,
+    WebhookLogResource,
+    WebhookLogListResource,
+    RegisterClientRequest,
+    UnRegisterClientRequest,
+    UnRegisterClientResponse,
+    ClientResource,
+    ClientListResource,
+    WebhookRegisteredOnResource,
+    WebhookRegisteredOnListResource,
 )
 
 __all__ = [
     "WebhookService",
-    "CreateServiceRequest",
-    "ServiceListResource",
     "ServiceResource",
-    "CreateWebhookRequest",
-    "UpdateWebhookRequest",
+    "ServiceListResource",
+    "CreateServiceRequest",
     "WebhookResource",
     "WebhookListResource",
-    "DeleteWebhookResponse",
+    "CreateWebhookRequest",
+    "EventResource",
     "EventListResource",
+    "UpdateWebhookRequest",
+    "DeleteWebhookResponse",
+    "WebhookLogResource",
     "WebhookLogListResource",
     "RegisterClientRequest",
     "UnRegisterClientRequest",
     "UnRegisterClientResponse",
+    "ClientResource",
     "ClientListResource",
+    "WebhookRegisteredOnResource",
     "WebhookRegisteredOnListResource",
 ]
