@@ -2,85 +2,102 @@
 Models for the Webhook service API.
 """
 from datetime import datetime
-from typing import List, Optional, Union, Dict, Any
+from enum import Enum
+from typing import List, Optional, Dict, Any
 
 from pydantic import BaseModel
 
 
-class ValidationError(BaseModel):
-    """Validation error model."""
-    loc: List[Union[str, int]]
-    msg: str
-    type: str
-
-
-class HTTPValidationError(BaseModel):
-    """HTTP validation error model."""
-    detail: List[ValidationError]
+class RequestMethodType(str, Enum):
+    POST = "POST"
+    PUT = "PUT"
+    PATCH = "PATCH"
 
 
 class ServiceResource(BaseModel):
     """Service resource model."""
-    id: int
-    name: str
+    id: Optional[int] = None
+    title: Optional[str] = None
     description: Optional[str] = None
-    created_at: str
-    updated_at: str
+    is_verified: Optional[bool] = None
+    is_active: Optional[bool] = None
+    created_at: Optional[str] = None
 
 
 class ServiceListResource(BaseModel):
     """Service list resource model."""
-    data: List[ServiceResource]
+    data: Optional[List[ServiceResource]] = None
+    result_count: Optional[int] = None
+    total_count: Optional[int] = None
+    total_page: Optional[int] = None
+    page: Optional[int] = None
+    per_page: Optional[int] = None
 
 
 class CreateServiceRequest(BaseModel):
     """Create service request model."""
-    name: str
-    description: Optional[str] = None
+    title: str
+    description: str
 
 
 class WebhookResource(BaseModel):
     """Webhook resource model."""
     id: int
     service_id: int
-    url: str
-    event_ids: List[str]
-    is_active: bool
-    created_at: str
-    updated_at: str
+    events: Optional[List[Dict[str, Any]]] = None
+    request_headers: Optional[Dict[str, Any]] = None
+    request_method: Optional[RequestMethodType] = None
+    url: Optional[str] = None
+    is_active: Optional[bool] = None
+    created_at: Optional[datetime] = None
+    updated_at: Optional[datetime] = None
 
 
 class WebhookListResource(BaseModel):
     """Webhook list resource model."""
-    data: List[WebhookResource]
+    data: Optional[List[WebhookResource]] = None
+    result_count: Optional[int] = None
+    total_count: Optional[int] = None
+    total_page: Optional[int] = None
+    page: Optional[int] = None
+    per_page: Optional[int] = None
 
 
 class CreateWebhookRequest(BaseModel):
     """Create webhook request model."""
-    service_id: int
+    service_id: Optional[int] = None
+    event_ids: List[int]
+    request_headers: Optional[str] = None
+    request_method: RequestMethodType
     url: str
-    event_ids: List[str]
-    is_active: bool = True
+    is_active: Optional[bool] = None
+    register_me: Optional[bool] = None
 
 
 class EventResource(BaseModel):
     """Event resource model."""
-    id: str
+    id: int
     name: str
     description: Optional[str] = None
-    service_id: int
+    sample_data: Optional[Dict[str, Any]] = None
+    scopes: Optional[str] = None
 
 
 class EventListResource(BaseModel):
     """Event list resource model."""
-    data: List[EventResource]
+    data: Optional[List[EventResource]] = None
+    result_count: Optional[int] = None
+    total_count: Optional[int] = None
+    total_page: Optional[int] = None
+    page: Optional[int] = None
+    per_page: Optional[int] = None
 
 
 class UpdateWebhookRequest(BaseModel):
     """Request model for updating a webhook."""
     event_ids: Optional[List[int]] = None
     request_headers: Optional[str] = None
-    request_method: Optional[str] = None
+    request_method: Optional[RequestMethodType] = None
     url: Optional[str] = None
     is_active: Optional[bool] = None
 
@@ -96,19 +113,19 @@ class WebhookLogResource(BaseModel):
     id: int
     user_id: int
     status_code: int
-    request: Optional[Dict[str, Any]] = None
-    response: Optional[Dict[str, Any]] = None
+    request: Dict[str, Any]
+    response: str
     created_at: Optional[datetime] = None
 
 
 class WebhookLogListResource(BaseModel):
     """Response model for list of webhook logs."""
     data: Optional[List[WebhookLogResource]] = None
-    result_count: Optional[int] = 1
+    result_count: Optional[int] = None
     total_count: Optional[int] = None
     total_page: Optional[int] = None
-    page: Optional[int] = 1
-    per_page: Optional[int] = 10
+    page: Optional[int] = None
+    per_page: Optional[int] = None
 
 
 class RegisterClientRequest(BaseModel):
@@ -141,11 +158,11 @@ class ClientResource(BaseModel):
 class ClientListResource(BaseModel):
     """Response model for list of clients."""
     data: Optional[List[ClientResource]] = None
-    result_count: Optional[int] = 1
+    result_count: Optional[int] = None
     total_count: Optional[int] = None
     total_page: Optional[int] = None
-    page: Optional[int] = 1
-    per_page: Optional[int] = 10
+    page: Optional[int] = None
+    per_page: Optional[int] = None
 
 
 class WebhookRegisteredOnResource(BaseModel):
@@ -153,16 +170,16 @@ class WebhookRegisteredOnResource(BaseModel):
     id: int
     service_id: int
     customer_id: int
-    events: Optional[List[Dict[str, Any]]] = None
-    is_active: Optional[bool] = True
+    events: List[Dict[str, Any]]
+    is_active: Optional[bool] = None
     registered_at: Optional[datetime] = None
 
 
 class WebhookRegisteredOnListResource(BaseModel):
     """Response model for list of webhook registrations."""
     data: Optional[List[WebhookRegisteredOnResource]] = None
-    result_count: Optional[int] = 1
+    result_count: Optional[int] = None
     total_count: Optional[int] = None
     total_page: Optional[int] = None
-    page: Optional[int] = 1
-    per_page: Optional[int] = 10
+    page: Optional[int] = None
+    per_page: Optional[int] = None
